mv client_update_ping_pong to tests dir
