+++ ibc-testkit/tests/core/ics02_client/update_client.rs
@@ -29,7 +29,10 @@ use ibc_testkit::fixtures::clients::tendermint::ClientStateConfig;
 use ibc_testkit::fixtures::core::context::MockContextConfig;
 use ibc_testkit::fixtures::core::signer::dummy_account_id;
 use ibc_testkit::hosts::tendermint::BlockParams;
-use ibc_testkit::hosts::{MockHost, TendermintHost, TestBlock, TestHeader, TestHost};
+use ibc_testkit::hosts::{
+    HostClientState, MockHost, TendermintHost, TestBlock, TestHeader, TestHost,
+};
+use ibc_testkit::relayer::error::RelayerError;
 use ibc_testkit::testapp::ibc::clients::mock::client_state::{
     client_type as mock_client_type, MockClientState,
 };
@@ -37,9 +40,12 @@ use ibc_testkit::testapp::ibc::clients::mock::header::MockHeader;
 use ibc_testkit::testapp::ibc::clients::mock::misbehaviour::Misbehaviour as MockMisbehaviour;
 use ibc_testkit::testapp::ibc::clients::AnyConsensusState;
 use ibc_testkit::testapp::ibc::core::router::MockRouter;
-use ibc_testkit::testapp::ibc::core::types::{LightClientBuilder, LightClientState, MockIbcStore};
+use ibc_testkit::testapp::ibc::core::types::{
+    DefaultIbcStore, LightClientBuilder, LightClientState, MockIbcStore,
+};
 use rstest::*;
 use tendermint_testgen::Validator as TestgenValidator;
+use tracing::debug;
 
 struct Fixture {
     ctx: MockContext<MockHost>,
@@ -1521,4 +1527,164 @@ fn test_client_update_max_clock_drift() {
 
     let res = validate(&ctx_a.ibc_store, &router_a, msg_envelope);
     assert!(res.is_err());
+}
+
+/// Builds a `ClientMsg::UpdateClient` for a client with id `client_id` running on the `dest`
+/// context, assuming that the latest header on the source context is `src_header`.
+pub(crate) fn build_client_update_datagram<H: TestHeader, Dst: TestHost>(
+    dest: &MockContext<Dst>,
+    client_id: &ClientId,
+    src_header: &H,
+) -> Result<ClientMsg, RelayerError>
+where
+    HostClientState<Dst>: ClientStateValidation<DefaultIbcStore>,
+{
+    // Check if client for ibc0 on ibc1 has been updated to latest height:
+    // - query client state on destination chain
+    let dest_client_latest_height = dest.light_client_latest_height(client_id);
+
+    if src_header.height() == dest_client_latest_height {
+        return Err(RelayerError::ClientAlreadyUpToDate {
+            client_id: client_id.clone(),
+            source_height: src_header.height(),
+            destination_height: dest_client_latest_height,
+        });
+    };
+
+    if dest_client_latest_height > src_header.height() {
+        return Err(RelayerError::ClientAtHigherHeight {
+            client_id: client_id.clone(),
+            source_height: src_header.height(),
+            destination_height: dest_client_latest_height,
+        });
+    };
+
+    // Client on destination chain can be updated.
+    Ok(ClientMsg::UpdateClient(MsgUpdateClient {
+        client_id: client_id.clone(),
+        client_message: src_header.clone().into(),
+        signer: dummy_account_id(),
+    }))
+}
+
+#[rstest]
+/// Serves to test both ICS-26 `dispatch` & `build_client_update_datagram` functions.
+/// Implements a "ping pong" of client update messages, so that two chains repeatedly
+/// process a client update message and update their height in succession.
+fn client_update_ping_pong() {
+    let chain_a_start_height = Height::new(1, 11).unwrap();
+    let chain_b_start_height = Height::new(1, 20).unwrap();
+    let client_on_b_for_a_height = Height::new(1, 10).unwrap(); // Should be smaller than `chain_a_start_height`
+    let client_on_a_for_b_height = Height::new(1, 20).unwrap(); // Should be smaller than `chain_b_start_height`
+    let num_iterations = 4;
+
+    let client_on_a_for_b = tm_client_type().build_client_id(0);
+    let client_on_b_for_a = mock_client_type().build_client_id(0);
+
+    let chain_id_a = ChainId::new("mockgaiaA-1").unwrap();
+    let chain_id_b = ChainId::new("mockgaiaB-1").unwrap();
+
+    // Create two mock contexts, one for each chain.
+    let mut ctx_a = MockContextConfig::builder()
+        .host(MockHost::builder().chain_id(chain_id_a).build())
+        .latest_height(chain_a_start_height)
+        .build::<MockContext<MockHost>>();
+
+    let mut ctx_b = MockContextConfig::builder()
+        .host(TendermintHost::builder().chain_id(chain_id_b).build())
+        .latest_height(chain_b_start_height)
+        .latest_timestamp(ctx_a.timestamp_at(chain_a_start_height.decrement().unwrap())) // chain B is running slower than chain A
+        .build::<MockContext<TendermintHost>>();
+
+    ctx_a = ctx_a.with_light_client(
+        &client_on_a_for_b,
+        LightClientBuilder::init()
+            .context(&ctx_b)
+            .consensus_heights([client_on_a_for_b_height])
+            .build(),
+    );
+
+    ctx_b = ctx_b.with_light_client(
+        &client_on_b_for_a,
+        LightClientBuilder::init()
+            .context(&ctx_a)
+            .consensus_heights([client_on_b_for_a_height])
+            .build(),
+    );
+
+    for _i in 0..num_iterations {
+        // Update client on chain B to latest height of A.
+        // - create the client update message with the latest header from A
+        let a_latest_header = ctx_a.query_latest_block().unwrap();
+        let client_msg_b_res = build_client_update_datagram(
+            &ctx_b,
+            &client_on_b_for_a,
+            &a_latest_header.into_header(),
+        );
+
+        assert!(
+                client_msg_b_res.is_ok(),
+                "create_client_update failed for context destination {ctx_b:?}, error: {client_msg_b_res:?}",
+            );
+
+        let client_msg_b = client_msg_b_res.unwrap();
+
+        // - send the message to B. We bypass ICS18 interface and call directly into
+        // MockContext `recv` method (to avoid additional serialization steps).
+        let dispatch_res_b = ctx_b.deliver(MsgEnvelope::Client(client_msg_b));
+        let validation_res = ctx_b.host.validate();
+        assert!(
+            validation_res.is_ok(),
+            "context validation failed with error {validation_res:?} for context {ctx_b:?}",
+        );
+
+        // Check if the update succeeded.
+        assert!(
+            dispatch_res_b.is_ok(),
+            "Dispatch failed for host chain b with error: {dispatch_res_b:?}"
+        );
+
+        assert_eq!(
+            ctx_b.light_client_latest_height(&client_on_b_for_a),
+            ctx_a.latest_height()
+        );
+
+        // Update client on chain A to latest height of B.
+        // - create the client update message with the latest header from B
+        // The test uses LightClientBlock that does not store the trusted height
+        let mut b_latest_header = ctx_b.query_latest_block().unwrap().clone().into_header();
+
+        let th = b_latest_header.height();
+        b_latest_header.set_trusted_height(th.decrement().unwrap());
+
+        let client_msg_a_res =
+            build_client_update_datagram(&ctx_a, &client_on_a_for_b, &b_latest_header);
+
+        assert!(
+                client_msg_a_res.is_ok(),
+                "create_client_update failed for context destination {ctx_a:?}, error: {client_msg_a_res:?}",
+            );
+
+        let client_msg_a = client_msg_a_res.unwrap();
+
+        debug!("client_msg_a = {:?}", client_msg_a);
+
+        // - send the message to A
+        let dispatch_res_a = ctx_a.deliver(MsgEnvelope::Client(client_msg_a));
+        let validation_res = ctx_a.host.validate();
+        assert!(
+            validation_res.is_ok(),
+            "context validation failed with error {validation_res:?} for context {ctx_a:?}",
+        );
+
+        // Check if the update succeeded.
+        assert!(
+            dispatch_res_a.is_ok(),
+            "Dispatch failed for host chain a with error: {dispatch_res_a:?}"
+        );
+        assert_eq!(
+            ctx_a.light_client_latest_height(&client_on_a_for_b),
+            ctx_b.latest_height()
+        );
+    }
 }
